requirements.txt
streamlit
pymupdf
langchain
cohere
faiss-cpu
openai
docx2txt
pillow
tenacity
tiktoken
thefuzz[speedup]
nlpcloud
sentence_transformers